requirements.txt
absl-py==2.1.0
adal==1.2.7
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.4.0
applicationinsights==0.11.10
argcomplete==3.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-appconfiguration==1.1.1
azure-batch==14.2.0
azure-cli==2.61.0
azure-cli-core==2.61.0
azure-cli-telemetry==1.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==31.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==30.0.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.4.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.0.0b6
azure-mgmt-servicelinker==1.2.0b2
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b16
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-query==1.2.0
azure-multiapi-storage==1.2.0
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.18.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bertopic==0.16.3
blinker==1.8.2
build==1.2.1
CacheControl==0.13.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.8
cycler==0.12.1
Cython==0.29.37
databricks-cli==0.18.0
dataclasses-json==0.6.7
deap==1.4.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fabric==3.2.2
Faker==29.0.0
fastjsonschema==2.19.1
filelock==3.13.3
findspark==2.0.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
hdbscan==0.8.37
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.6
immutabledict==4.2.0
importlib_metadata==7.1.0
installer==0.7.0
invoke==2.2.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.4
isodate==0.6.1
jaraco.classes==3.4.0
javaproperties==0.5.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsondiff==2.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.12
keyring==24.3.1
kiwisolver==1.4.5
knack==0.11.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langchainhub==0.1.18
langsmith==0.1.101
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.2.0b1
msgpack==1.0.8
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.23.1
nest-asyncio==1.6.0
networkx==3.3
ngrok==1.4.0
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.33.0
openpyxl==3.1.4
orjson==3.10.7
ortools==9.10.4067
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==3.11.0
plotly==5.23.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pyarrow==16.1.0
pycomposefile==0.0.31
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pymsalruntime==0.14.2
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.7.0
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
scp==0.13.6
seaborn==0.13.2
semver==2.13.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sshtunnel==0.1.5
stack-data==0.6.3
streamlit==1.35.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
torch==2.4.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.43.3
trove-classifiers==2024.3.25
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
websocket-client==1.3.3
widgetsnbextension==4.0.12
wikipedia==1.4.0
wordcloud==1.9.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
yfiles_jupyter_graphs==1.7.3
zipp==3.18.1
